Now displays (dummy) estimated distance on screen

+++ vehicle_detection/DistanceEstimation.h
@@ -6,7 +6,7 @@ using namespace std;
 
 void initializeDistanceEstimation();
 
-double estimateDistance(float camXCoord, float camYCoord);
+float estimateDistance(float camXCoord, float camYCoord);
 
 
 
@@ -15,6 +15,6 @@ void initializeDistanceEstimation() {
 }
 
 
-double estimateDistance(float camXCoord, float camYCoord) {
+float estimateDistance(float camXCoord, float camYCoord) {
 	return 3.14;
 }
